+++ trace.c
@@ -52,9 +52,9 @@ int print_trace(int mode, char *fmt, ...) {
     }
 
     fprintf(log_file, "TCF %02d:%02d.%03d: %s\n",
-        timenow.tv_sec / 60 % 60,
-        timenow.tv_sec % 60,
-        timenow.tv_nsec / 1000000,
+        (int)(timenow.tv_sec / 60 % 60),
+        (int)(timenow.tv_sec % 60),
+        (int)(timenow.tv_nsec / 1000000),
         tmpbuf);
     fflush(log_file);
